Fix bug in varm (SSE2 version) associated with unaligned input data
In the case of unaligned input data, n is decremented up to three times so that the majority of computations can be conducted with aligned data. As the original value of n is needed for the final division, it has to be saved in a separate variable to avoid erroneously using the decremented value there.